refactor: Notification Channel 통합
